Small changes from abandoned light matter reader branch.
