+++ dark/blast/alignments.py
@@ -1,47 +1,15 @@
-from os.path import basename
 from random import uniform
 from math import log10
 import copy
-import string
 
 from dark.score import HigherIsBetterScore
 from dark.alignments import ReadsAlignments, ReadsAlignmentsParams
 from dark.blast.conversion import JSONRecordsReader
 from dark.blast.params import checkCompatibleParams
 from dark import ncbidb
+from dark.utils import numericallySortFilenames
 
 ZERO_EVALUE_UPPER_RANDOM_INCREMENT = 150
-
-
-def numericallySortFilenames(names):
-    """
-    Sort (ascending) a list of file names by their numerical prefixes.
-    The number sorted on is the numeric prefix of the basename of
-    the given filename. E.g., '../output/1.json.bz2' will sort before
-    '../output/10.json.bz2'.
-
-    @param: A C{list} of file names, each of whose basename starts with a
-        string of digits.
-    @return: The sorted C{list} of full file names.
-    """
-
-    def numericPrefix(name):
-        """
-        Find any numeric prefix of C{name} and return it as an C{int}.
-
-        @param: A C{str} file name, whose name possibly starts with digits.
-        @return: The C{int} number at the start of the name, else 0 if
-            there are no leading digits in the name.
-        """
-        count = 0
-        for ch in name:
-            if ch in string.digits:
-                count += 1
-            else:
-                break
-        return 0 if count == 0 else int(name[0:count])
-
-    return sorted(names, key=lambda name: numericPrefix(basename(name)))
 
 
 class BlastReadsAlignments(ReadsAlignments):
@@ -157,9 +125,9 @@ class BlastReadsAlignments(ReadsAlignments):
                 'records found (%d). First unknown read id is %r.' %
                 (count, read.id))
 
-    def getSequence(self, title):
+    def getSubjectSequence(self, title):
         """
-        Obtain information about a sequence, given its title.
+        Obtain information about a subject sequence, given its title.
 
         @param title: A C{str} sequence title from a BLAST hit. Of the form
             'gi|63148399|gb|DQ011818.1| Description...'.
